pyproject.toml
[project]
name = "jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-allauth>=65.12.0",
    "djangorestframework>=3.16.1",
]
